+++ pyfar/__init__.py
@@ -11,7 +11,6 @@ from .classes.audio import Signal, TimeData, FrequencyData
 from .classes.audio import (add, subtract, multiply, divide, power,
                             matrix_multiplication)
 from .classes.coordinates import Coordinates
-from .classes.coordinates import SamplingSphere
 from .classes.orientations import Orientations
 from .classes.filter import FilterFIR, FilterIIR, FilterSOS
